+++ bin/cogs/Abusing.py
@@ -1,10 +1,11 @@
 import re
 
-from bin.utils.channels import get_channel
-from db import db
 from discord import Embed
 from discord.ext.commands import Cog
+
 from resources.data.regex import REGEX
+from .. import db
+from ..utils import get_channel
 
 
 class Abusing(Cog):
@@ -15,8 +16,7 @@ class Abusing(Cog):
     async def on_message(self, message):
         """Antimat system."""
         if not message.author.bot:
-            status = await db.record('SELECT abuse FROM configs WHERE '
-                                     'guild_id = %s', message.guild.id)
+            status = await db.record('SELECT abuse FROM configs WHERE guild_id = %s', message.guild.id)
 
             if status == 'on':
                 result = re.findall(REGEX, message.content)