+++ bin/cogs/GuildAdding.py
@@ -1,7 +1,6 @@
 from discord.ext.commands import Cog, command, is_owner
-from loguru import logger
 
-from db import db
+from .. import db
 
 
 class GuildAdding(Cog):
@@ -12,40 +11,28 @@ class GuildAdding(Cog):
     @is_owner()
     async def add_guild_to_db(self, ctx):
         """Ручное добавление сервера в бд"""
-        results = await db.field('SELECT * FROM configs WHERE guild_id = %s',
-                                 ctx.guild.id)
+        results = await db.record('SELECT * FROM configs WHERE guild_id = %s', ctx.guild.id)
 
         if not results:
-            await db.execute('INSERT INTO configs (guild_id, user_log, guest,'
-                             ' abuse, welcoming) VALUES(%s, %s, %s, %s)',
+            await db.execute('INSERT INTO configs (guild_id, user_log, guest, abuse, welcoming) VALUES(%s, %s, %s, %s)',
                              ctx.guild.id, 'off', 'off', 'off', 'off')
-            await db.execute('INSERT INTO channels (guild_id) VALUES(%s)',
-                             ctx.guild.id)
-            await db.execute('INSERT INTO roles (guild_id) VALUES(%s)',
-                             ctx.guild.id)
+            await db.execute('INSERT INTO channels (guild_id) VALUES(%s)', ctx.guild.id)
+            await db.execute('INSERT INTO roles (guild_id) VALUES(%s)', ctx.guild.id)
             await db.commit()
         elif results:
-            await ctx.send(f'{ctx.message.author.mention}, this guild is '
-                           'already exist in DB.')
+            await ctx.send(f'{ctx.message.author.mention}, данный сервер уже зарегестрирован в базе данных бота.')
 
     @Cog.listener()
     async def on_guild_join(self, guild):
         """Автоматическое добавление сервера в бд"""
-        results = await db.field('SELECT * FROM configs WHERE guild_id = %s',
-                                 guild.id)
+        results = await db.record('SELECT * FROM configs WHERE guild_id = %s', guild.id)
 
         if not results:
-            await db.execute('INSERT INTO configs (guild_id, user_log, guest,'
-                             ' abuse, welcoming) VALUES(%s, %s, %s, %s)',
+            await db.execute('INSERT INTO configs (guild_id, user_log, guest, abuse, welcoming) VALUES(%s, %s, %s, %s)',
                              guild.id, 'off', 'off', 'off', 'off')
-            await db.execute('INSERT INTO channels (guild_id) VALUES(%s)',
-                             guild.id)
-            await db.execute('INSERT INTO roles (guild_id) VALUES(%s)',
-                             guild.id)
+            await db.execute('INSERT INTO channels (guild_id) VALUES(%s)', guild.id)
+            await db.execute('INSERT INTO roles (guild_id) VALUES(%s)', guild.id)
             await db.commit()
-        elif results:
-            logger.info('Guild is already exist in DB. ID: {guild.id}, '
-                        'Name: {guild.name}')
 
 
 def setup(bot):