+++ bin/cogs/MessageLogs.py
@@ -1,8 +1,9 @@
-from bin.utils.channels import get_channel
 from discord import Embed
+from discord.errors import HTTPException
 from discord.ext.commands import Cog
 
 from .. import db
+from ..utils.channels import get_channel
 
 
 class MessageLogs(Cog):
@@ -11,8 +12,7 @@ class MessageLogs(Cog):
 
     @Cog.listener()
     async def on_raw_message_delete(self, payload):
-        status = await db.record('SELECT logging FROM configs WHERE '
-                                 'guild_id = %s', payload.guild_id)
+        status = await db.record('SELECT logging FROM configs WHERE guild_id = %s', payload.guild_id)
 
         if status == 'on':
             message = payload.cached_message
@@ -28,15 +28,17 @@ class MessageLogs(Cog):
                           ('Канал', msg_channel.mention, True)]
 
                 if message.embeds:
-                    fields.append(('Количество встроенных врезок',
-                                   len(message.embeds), True))
+                    fields.append(('Количество встроенных врезок', len(message.embeds), True))
                 else:
                     if message.attachments:
                         if message.content:
                             fields.append(('Контент', message.content, False))
 
                         image_url = message.attachments[0].proxy_url == ''
-                        embed.set_image(url=image_url)
+                        try:
+                            embed.set_image(url=image_url)
+                        except HTTPException:
+                            fields.append(('Изображение', 'Изображение не найдено в кэше', True))
                     else:
                         fields.append(('Контент', message.content, False))
 
@@ -47,8 +49,7 @@ class MessageLogs(Cog):
 
     @Cog.listener()
     async def on_raw_bulk_message_delete(self, payload):
-        status = await db.record('SELECT logging FROM configs WHERE '
-                                 'guild_id = %s', payload.guild_id)
+        status = await db.record('SELECT logging FROM configs WHERE guild_id = %s', payload.guild_id)
 
         if status == 'on':
             messages = payload.cached_messages
@@ -57,12 +58,10 @@ class MessageLogs(Cog):
                 log_channel = await get_channel(payload.guild_id)
                 msg_channel = self.bot.get_channel(payload.channel_id)
 
-                embed = Embed(title='Группа сообщений была удалена',
-                              color=0xFF5733)
+                embed = Embed(title='Группа сообщений была удалена', color=0xFF5733)
 
                 fields = [('Канал', msg_channel.mention, True),
-                          ('Количество сообщений',
-                           len(payload.cached_messages), True)]
+                          ('Количество сообщений', len(payload.cached_messages), True)]
 
                 for name, value, inline in fields:
                     embed.add_field(name=name, value=value, inline=inline)