+++ bin/db/db.py
@@ -1,4 +1,5 @@
 import psycopg2
+
 from resources.data.tokens.db_keys import dbHost, dbName, dbPass, dbUser
 
 conn = psycopg2.connect(