+++ bin/utils/__init__.py
@@ -1,0 +1,4 @@
+from .channels import *
+from .dates import *
+from .roles import *
+from .timetables import *