+++ bin/utils/roles.py
@@ -1,8 +1,10 @@
-import db
-from src import bot
+from .. import db
+from ..src import bot
 
 
 async def get_role(guild_id: int, role: str):
+    """Функция получает ID роли из БД
+    Доступные роли: `guest_role`"""
     guild = bot.get_guild(guild_id)
 
     role_id = await db.record(f'SELECT {role} FROM roles WHERE guild_id = %s',
@@ -11,15 +13,3 @@ async def get_role(guild_id: int, role: str):
         role = guild.get_role(role_id)
 
         return role
-
-
-async def get_guest_role(guild_id: int):
-    guest_role = await get_role(guild_id, 'guest_role')
-
-    return guest_role
-
-
-async def get_curator_role(guild_id: int):
-    curator_role = await get_role(guild_id, 'curator_role')
-
-    return curator_role