+++ CHANGELOG.md
@@ -1,100 +1,129 @@
 # Changelog
 
+## v0.1.20 – 2020-12-18
+
+**Updated:**
+
+- Fix imports
+- Random member choice update
+
 ## v0.1.19 – 2020-12-09
+
 **Added:**
+
 - Choice random member from role command
 
 **Updated:**
+
 - flake8 fix
 
-
 ## v0.1.18 – 2020-11-23
+
 **Added:**
+
 - Direct welcoming messages
 
 **Updated:**
+
 - Some changes
 
-
 ## v0.1.17 – 2020-11-21
+
 **Added:**
+
 - Create raw message delete
 - Create raw bulk message delete
 
 **Updated:**
+
 - New license
 - Some syntax upds
 
-
 ## v0.1.16 – 2020-11-19
-- Checked issues by flake8
 
+- Checked issues by flake8
 
 ## v0.1.15 – 2020-11-19
+
 - Checked issues by flake8
 
-
 ## v0.1.14 – 2020-11-14
+
 **Added:**
+
 - Help command
 - Command description
 
 ## v0.1.13 – 2020-11-10
+
 **Updated:**
+
 - Updates hierarchy of files
 - Some utils updates
 
-
 ## v0.1.12 – 2020-11-05
+
 **Added:**
+
 - Add kick command
 - Add ban command
 
 **Updated:**
+
 - Little upds
 
-
 ## v0.1.10 – 2020-10-31
+
 **Added:**
+
 - Timetable for next week
 
 **Updated:**
+
 - Added action time
 
-
 ## v0.1.7 – 2020-10-28
+
 **Added:**
+
 - Get number of week in year
 
 **Updated:**
+
 - Some updates in get timetable from GSheets
 
-
 ## v0.1.6 – 2020-10-25
+
 **Added:**
+
 - Curator role
 - Curator settings command
 
 **Updated:**
+
 - Update DB functions
 
-
 ## v0.1.5 – 2020-10-25
+
 **Added:**
+
 - Clear chat command
 
-
 ## v0.1.4 – 2020-10-24
+
 **Added:**
+
 - More errors handler added
 - Create Statistics
 
 **Updated:**
+
 - Utils update
 
-
 ## v0.1.3 – 2020-10-23
+
 **Added:**
+
 - Timetables
 - Create ErrorHandler file
 - Create TrafficHandler file
@@ -104,9 +133,11 @@
 - License
 
 **Updated:**
+
 - Abusing system
-  - Some regex upds
+    - Some regex upds
 - Member events
 
 **Bug fixes:**
+
 - Get guild owners command